requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
google-auth==1.6.3
idna==2.8
kubernetes==9.0.0
multidict==4.5.2
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycares==3.0.0
pycparser==2.19
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
slackclient==2.0.0
urllib3==1.25.2
websocket-client==0.56.0
yarl==1.3.0
